moved chain files into server
